fix: Round float-to-integer casts instead of truncating (#2639)
MySQL's CAST to SIGNED/INTEGER rounds to the nearest integer rather
than truncating toward zero. This fixes CAST(-0.7 AS SIGNED) to
correctly return -1 instead of 0.

Changes:
- Update cast_value() in casting.rs to use .round() for Numeric, Float,
  Real, and Double conversions to integer types (Integer, Smallint,
  Bigint, Unsigned)
- Update to_i64() helper function to use rounding
- Add comprehensive tests for rounding edge cases
- Update existing tests that incorrectly expected truncation

Closes #2639

+++ crates/vibesql-executor/tests/conversion_function_tests.rs
@@ -397,8 +397,9 @@ fn test_cast_integer_to_double() {
 
 #[test]
 fn test_cast_double_to_integer() {
+    // MySQL rounds to nearest integer when casting to INTEGER (not truncate)
     let result = eval_function("CAST", vec![double_lit(42.9), varchar_lit("INTEGER")]);
-    assert_eq!(result, vibesql_types::SqlValue::Integer(42));
+    assert_eq!(result, vibesql_types::SqlValue::Integer(43));
 }
 
 #[test]
